Refactor: entity 패키지 이름을 domain으로 변경, CSV 파일 경로를 상수화, Scanner 상수 이름을 ConsoleInput으로 수정한다.
