+++ src/main/java/common/Util.java
@@ -3,5 +3,5 @@ package common;
 import java.util.Scanner;
 
 public class Util {
-    public static Scanner sc = new Scanner(System.in);
+    public static Scanner ConsoleInput = new Scanner(System.in);
 }
