+++ src/main/java/domain/Book.java
@@ -1,4 +1,4 @@
-package entity;
+package domain;
 
 import java.util.List;
